Use long for dates

+++ AppStoreServerApi/Models/JWSRenewalInfoDecodedPayload.cs
@@ -12,7 +12,7 @@ namespace AppStoreServerApi.Models
         public string AutoRenewProductId { get; set; } = null!;
         public AutoRenewStatus AutoRenewStatus { get; set; }
         public ExpirationIntent? ExpirationIntent { get; set; } = null!;
-        public int? GracePeriodExpiresDate { get; set; } = null!;
+        public long? GracePeriodExpiresDate { get; set; } = null!;
         public bool? IsInBillingRetryPeriod { get; set; } = null!;
         public string? OfferIdentifier { get; set; } = null!;
         public OfferType? OfferType { get; set; } = null!;
@@ -22,6 +22,6 @@ namespace AppStoreServerApi.Models
         /// <summary>
         /// UNIX timestamp in ms
         /// </summary>
-        public int SignedDate { get; set; }
+        public long SignedDate { get; set; }
     }
 }
